commit report

+++ src/test/java/utilities/CapReport.java
@@ -2,17 +2,17 @@ package utilities;
 
 public class CapReport {
 	
-	private float subscription;
+	private int subscription;
 	private String client;
 	private String user;
-	private float numberOrder;
-	private float totalLead;
-	private float totalCap;
-	private float missedLead;
-	private float difference;
+	private int numberOrder;
+	private int totalLead;
+	private int totalCap;
+	private int missedLead;
+	private int difference;
 	
-	public CapReport(float subscription, String client, String user, float numberOrder, float totalLead, float totalCap,
-			float missedLead, float difference) {
+	public CapReport(int subscription, String client, String user, int numberOrder, int totalLead, int totalCap,
+			int missedLead, int difference) {
 		this.subscription = subscription;
 		this.client = client;
 		this.user = user;
@@ -22,10 +22,10 @@ public class CapReport {
 		this.missedLead = missedLead;
 		this.difference = difference;
 	}
-	public float getSubscription() {
+	public int getSubscription() {
 		return subscription;
 	}
-	public void setSubscription(float subscription) {
+	public void setSubscription(int subscription) {
 		this.subscription = subscription;
 	}
 	public String getClient() {
@@ -34,34 +34,34 @@ public class CapReport {
 	public void setClient(String client) {
 		this.client = client;
 	}
-	public float getTotalLead() {
+	public int getTotalLead() {
 		return totalLead;
 	}
-	public void setTotalLead(float totalLead) {
+	public void setTotalLead(int totalLead) {
 		this.totalLead = totalLead;
 	}
-	public float getTotalCap() {
+	public int getTotalCap() {
 		return totalCap;
 	}
-	public void setTotalCap(float totalCap) {
+	public void setTotalCap(int totalCap) {
 		this.totalCap = totalCap;
 	}
-	public float getMissedLead() {
+	public int getMissedLead() {
 		return missedLead;
 	}
-	public void setMissedLead(float missedLead) {
+	public void setMissedLead(int missedLead) {
 		this.missedLead = missedLead;
 	}
-	public float getDifference() {
+	public int getDifference() {
 		return difference;
 	}
-	public void setDifference(float difference) {
+	public void setDifference(int difference) {
 		this.difference = difference;
 	}
-	public float getNumberOrder() {
+	public int getNumberOrder() {
 		return numberOrder;
 	}
-	public void setNumberOrder(float numberOrder) {
+	public void setNumberOrder(int numberOrder) {
 		this.numberOrder = numberOrder;
 	}
 	public String getUser() {
